Expanded wrapped vs. ported GDX object tests to run batch benchmarks over all suites

+++ tests/gdlib/gmsheapnewtests.cpp
@@ -22,13 +22,13 @@ namespace tests::gmsheapnew {
         std::array<uint8_t, 20> buf2{};
         int64_t *ref {(int64_t*)buf2.data()};
         for(int i{}; i<n; i++) {
-            ptrs[i] = thm.XGetMem(buf2.size());
+            ptrs[i] = thm.XGetMem((int)buf2.size());
             *ref = (int64_t)i;
             std::memcpy(ptrs[i], buf2.data(), buf2.size());
         }
         for(int i{}; i<n; i++) {
             REQUIRE_EQ(*((int64_t*)ptrs[i]), (int64_t)i);
-            thm.XFreeMem(ptrs[i], buf2.size());
+            thm.XFreeMem(ptrs[i], (int)buf2.size());
         }
     }
